requirements.txt
aiohttp
aiozmq
aiohttp_cors
msgpack-python
swiglpk==1.2.14
potion_client
requests
redis
codecov
gunicorn
uvloop
pytest-asyncio==0.5.0
pytest-cov==2.4.0